backend/requirements.txt
anyio==4.9.0
boto3==1.37.27
botocore==1.37.27
certifi==2025.1.31
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
jmespath==1.0.1
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.4.8
pydantic==1.10.13
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.29.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
